+++ next.config.js
@@ -3,6 +3,18 @@ module.exports = {
   future: {
     webpack5: true,
   },
+  async redirects() {
+    return [
+      {
+        source: "/logout",
+        destination: "/api/auth/logout",
+        permanent: true,
+      },
+    ];
+  },
+  images: {
+    domains: ["community.cloudflare.steamstatic.com"],
+  },
   webpack: (config, { dev, isServer }) => {
     if (!dev && !isServer) {
       config.resolve.alias = {