+++ src/pages/_document.tsx
@@ -25,6 +25,11 @@ class SnailyCAD extends Document {
             integrity="sha384-eOJMYsd53ii+scO/bJGFsiCZc+5NDVN2yr8+0RDqr0Ql0h+rP48ckxlpbzKgwra6"
             crossOrigin="anonymous"
           />
+          <script
+            src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.0-beta3/dist/js/bootstrap.bundle.min.js"
+            integrity="sha384-JEW9xMcG8R+pH31jmWH6WWP0WintQrMb4s7ZOdauHnUtxwoG2vI5DkLtS3qm9Ekf"
+            crossOrigin="anonymous"
+          ></script>
         </Head>
         <body>
           <Main />